[improvement](jdbc catalog) Optimize JdbcCatalog case mapping stability
