+++ fe/fe-core/src/main/java/org/apache/doris/datasource/ExternalCatalog.java
@@ -45,6 +45,7 @@ import org.apache.doris.datasource.iceberg.IcebergExternalDatabase;
 import org.apache.doris.datasource.infoschema.ExternalInfoSchemaDatabase;
 import org.apache.doris.datasource.infoschema.ExternalMysqlDatabase;
 import org.apache.doris.datasource.jdbc.JdbcExternalDatabase;
+import org.apache.doris.datasource.mapping.IdentifierMapping;
 import org.apache.doris.datasource.maxcompute.MaxComputeExternalDatabase;
 import org.apache.doris.datasource.metacache.MetaCache;
 import org.apache.doris.datasource.operations.ExternalMetadataOps;
@@ -142,6 +143,9 @@ public abstract class ExternalCatalog
     protected Optional<Boolean> useMetaCache = Optional.empty();
     protected MetaCache<ExternalDatabase<? extends ExternalTable>> metaCache;
 
+    protected IdentifierMapping identifierMapping;
+    private boolean mappingsInitialized = false;
+
     public ExternalCatalog() {
     }
 
@@ -174,6 +178,10 @@ public abstract class ExternalCatalog
         }
     }
 
+    // only for forward to master
+    protected void buildDatabaseMapping() {
+    }
+
     // Will be called when creating catalog(so when as replaying)
     // to add some default properties if missing.
     public void setDefaultPropsIfMissing(boolean isReplay) {
@@ -201,6 +209,10 @@ public abstract class ExternalCatalog
      * @return names of tables in specified database
      */
     public abstract List<String> listTableNames(SessionContext ctx, String dbName);
+
+    // only for forward to master
+    protected void buildTableMapping(SessionContext ctx, String dbName) {
+    }
 
     /**
      * check if the specified table exist.
@@ -265,6 +277,10 @@ public abstract class ExternalCatalog
                 init();
             }
             initialized = true;
+        }
+        if (!mappingsInitialized) {
+            buildDatabaseMapping();
+            mappingsInitialized = true;
         }
     }
 
@@ -391,6 +407,7 @@ public abstract class ExternalCatalog
     public void onRefresh(boolean invalidCache) {
         this.objectCreated = false;
         this.initialized = false;
+        this.mappingsInitialized = false;
         synchronized (this.propLock) {
             this.convertedProperties = null;
         }
@@ -716,6 +733,7 @@ public abstract class ExternalCatalog
         }
         this.propLock = new byte[0];
         this.initialized = false;
+        this.mappingsInitialized = false;
         setDefaultPropsIfMissing(true);
     }
