+++ fe/fe-core/src/main/java/org/apache/doris/datasource/ExternalDatabase.java
@@ -91,6 +91,8 @@ public abstract class ExternalDatabase<T extends ExternalTable>
 
     private MetaCache<T> metaCache;
 
+    private boolean mappingsInitialized = false;
+
     /**
      * Create external database.
      *
@@ -117,6 +119,7 @@ public abstract class ExternalDatabase<T extends ExternalTable>
 
     public void setUnInitialized(boolean invalidCache) {
         this.initialized = false;
+        this.mappingsInitialized = false;
         this.invalidCacheInInit = invalidCache;
         if (extCatalog.getUseMetaCache().isPresent()) {
             if (extCatalog.getUseMetaCache().get() && metaCache != null) {
@@ -169,6 +172,10 @@ public abstract class ExternalDatabase<T extends ExternalTable>
                 init();
             }
             initialized = true;
+        }
+        if (!mappingsInitialized) {
+            extCatalog.buildTableMapping(null, name);
+            mappingsInitialized = true;
         }
     }
